Adaptive stepping in subsims only
Norma supports different time steps for
each subdomain in a multi-domain simulation,
thus each sub-simulation can reduce or increase
its time step independently within a controller
stepper step, leaving the global step unchanged.

There may be a need to change the global step,
for example in contact, to capture contact
intitiation and release times, but the mechanism
will be different and automatic, with no need
for user parameters such as min and max time
steps or increase and decrease factors.